+++ packages/blob-stores/fileSystemBlobStore.js
@@ -31,7 +31,7 @@ const MAX_COUNTER = Math.pow(2, 31)
 class FileSystemBlobStore extends BlobStoreCustom {
   /**
    * @param {Object} storeConfig
-   * @param {App} appInstance
+   * @param {ServerApp} appInstance
    * @param {UBSession} sessionInstance
    */
   constructor (storeConfig, appInstance, sessionInstance) {