+++ packages/ub/modules/App.js
@@ -23,10 +23,10 @@ if (base.ubVersionNum < 5018000) {
  *
  *  - `launchEndpoint:before` with parameters: (req, resp, endpointName)
  *  - `endpointName + ':before'` event before endpoint handler  execution
- *  - `endpointName + ':after'` event in case neither exception is raised nor App.preventDefault()
+ *  - `endpointName + ':after'` event in case neither exception is raised nor App.preventDefault() is called
  *  - `launchEndpoint:after` with parameters: (req, resp, endpointName, defaultPrevented)
  *
- * To prevent endpoint handler execution App.preventDefault() can be used inside `:before` handler.
+ * To prevent endpoint handler execution `App.preventDefault()` can be used inside `:before` handler.
  *
  * @example
 const App = require('@unitybase/ub').App
@@ -34,8 +34,8 @@ const App = require('@unitybase/ub').App
 App.registerEndpoint('echoToFile', echoToFile, false)
 
 // write custom request body to file FIXTURES/req and echo file back to client
-// @param {THTTPRequest} req
-// @param {THTTPResponse} resp
+// @ param {THTTPRequest} req
+// @ param {THTTPResponse} resp
 function echoToFile (req, resp) {
   var fs = require('fs')
   fs.writeFileSync(path.join(FIXTURES, 'req'), req.read('bin'))
@@ -44,8 +44,8 @@ function echoToFile (req, resp) {
 }
 
 //Before getDocument requests
-//@param {THTTPRequest} req
-//@param {THTTPResponse} resp
+//@ param {THTTPRequest} req
+//@ param {THTTPResponse} resp
 function doSomethingBeforeGetDocumentCall(req, resp){
   console.log('User with ID', Session.userID, 'try to get document')
 }
@@ -54,27 +54,30 @@ App.on('getDocument:before', doSomethingBeforeGetDocumentCall)
 
 //
 //After getDocument requests
-//@param {THTTPRequest} req
-//@param {THTTPResponse} resp
+//@ param {THTTPRequest} req
+//@ param {THTTPResponse} resp
 function doSomethingAfterGetDocumentCall(req, resp){
   params = req.parsedParameters
   console.log('User with ID', Session.userID, 'obtain document using params',  params)
 }
 App.on('getDocument:after', doSomethingAfterGetDocumentCall)
  *
- * @class App
+ * @class ServerApp
  * @mixes EventEmitter
  */
-const App = {
-  /**
-   * Fires for an {@link App App} just after all domain entities (all *.meta) are loaded into server memory
-   * and all server-side js are evaluated.
-   *
-   * On this stage you can subscribe on a cross-model handles.
-   *
-   * @example:
-   *
-   App.once('domainIsLoaded', function(){
+const ServerApp = {}
+
+/**
+ * Fires for an {@link module:@unitybase/ub#App UB.App} just after all domain entities (all *.meta) are loaded into server memory
+ * and all server-side js are evaluated.
+ *
+ * On this stage you can subscribe on a cross-model handles.
+ *
+ * @example:
+ *
+ const UB = require('@unitybase/ub')
+ const App = UB.App
+ App.once('domainIsLoaded', function(){
      for (eName in App.domainInfo.entities) {
         // if entity have attribute mi_fedUnit
         if (App.domainInfo.entities[eName].attributes.mi_fedUnit) {
@@ -83,14 +86,24 @@ const App = {
         }
      }
    })
-   *
-   * @event domainIsLoaded
-   */
-}
+ *
+ * @event domainIsLoaded
+ * @memberOf ServerApp
+ */
+
+/**
+ * Fires (by native code) for an {@link module:@unitybase/ub#App UB.App} just after HTTP request context tries to got a DB connection for the first time
+ *
+ * On this stage DB session properties, specific for a current Session can be sets.
+ * For example multitenancy mixin subscribes for this event and sets a `ub_tenantID` DB session variable value to `Session.tenantID`
+ *
+ * @event enterConnectionContext
+ * @memberOf ServerApp
+ */
 
 // add eventEmitter to application object
-EventEmitter.call(App)
-Object.assign(App, EventEmitter.prototype)
+EventEmitter.call(ServerApp)
+Object.assign(ServerApp, EventEmitter.prototype)
 
 let __preventDefault = false
 /**
@@ -102,7 +115,7 @@ let __preventDefault = false
  * @private
  * @returns {boolean}
  */
-App.emitWithPrevent = function (eventName, req, resp) {
+ServerApp.emitWithPrevent = function (eventName, req, resp) {
   __preventDefault = false
   this.emit(eventName, req, resp)
   return __preventDefault
@@ -110,9 +123,9 @@ App.emitWithPrevent = function (eventName, req, resp) {
 /**
  * Accessible inside app-level `:before` event handler. Call to prevent default method handler.
  * In this case developer are responsible to fill response object, otherwise HTTP 400 is returned.
- * @memberOf App
+ * @memberOf ServerApp
  */
-App.preventDefault = function () {
+ServerApp.preventDefault = function () {
   __preventDefault = true
 }
 
@@ -124,12 +137,36 @@ App.preventDefault = function () {
  * @private
  * @returns {boolean}
  */
-App.launchEndpoint = function (endpointName) {
+ServerApp.launchEndpoint = function (endpointName) {
   __preventDefault = false
   const req = new THTTPRequest()
   const resp = new THTTPResponse()
   try {
+    /**
+     * Fires before any endpoint execution
+     * @event launchEndpoint:before
+     * @memberOf ServerApp
+     * @param {THTTPRequest} req
+     * @param {THTTPResponse} resp
+     * @param {string} endpointName
+     */
     this.emit('launchEndpoint:before', req, resp, endpointName)
+    /**
+     * Fires before endpoint execution. In example below handler is called before each `getDocument` execution
+     * @example
+const UB = require('@unitybase/ub')
+const App = UB.App
+function doSomethingBeforeGetDocumentCall(req, resp){
+  console.log('User with ID', Session.userID, 'try to get document')
+}
+// Adds hook called before each call to getDocument endpoint
+App.on('getDocument:before', doSomethingBeforeGetDocumentCall)
+     * @event endpointName:before
+     * @memberOf ServerApp
+     * @param {THTTPRequest} req
+     * @param {THTTPResponse} resp
+     * @param {string} endpointName
+     */
     this.emit(endpointName + ':before', req, resp)
     if (!__preventDefault) {
       const handler = appBinding.endpoints[endpointName]
@@ -138,11 +175,27 @@ App.launchEndpoint = function (endpointName) {
       } else { // native endpoint
         appBinding.launchNativeEndpoint()
       }
+      /**
+       * Fires after endpoint execution. In example below handler is called before each `getDocument` execution
+       * @event endpointName:after
+       * @memberOf ServerApp
+       * @param {THTTPRequest} req
+       * @param {THTTPResponse} resp
+       * @param {string} endpointName
+       */
       this.emit(endpointName + ':after', req, resp)
     }
+    /**
+     * Fires after any endpoint execution
+     * @event launchEndpoint:before
+     * @memberOf ServerApp
+     * @param {THTTPRequest} req
+     * @param {THTTPResponse} resp
+     * @param {string} endpointName
+     */
     this.emit('launchEndpoint:after', req, resp, endpointName, __preventDefault)
   } finally {
-    App.endpointContext = {} // allow GC to release possible context data ASAP
+    ServerApp.endpointContext = {} // allow GC to release possible context data ASAP
   }
 }
 
@@ -151,7 +204,7 @@ App.launchEndpoint = function (endpointName) {
  * @param {ubMethodParams} ctxt
  * @private
  */
-App.launchRLS = function (ctxt) {
+ServerApp.launchRLS = function (ctxt) {
   const rlsMixin = ctxt.dataStore.entity.mixins.rls
   if (rlsMixin.func) { // functional RLS
     if (!rlsMixin.__funcVar) { // parse func namespace 'uba_user.somefunc' (string) -> uba_user.somefunc (function)
@@ -214,9 +267,9 @@ const appBinding = process.binding('ub_app')
  * @param {boolean} [authorizationRequired=true] If `true` UB will check for valid Authorization header before
  *  execute endpoint handler
  * @param {boolean} [isDefault=false]
- * @memberOf App
+ * @memberOf ServerApp
  */
-App.registerEndpoint = function (endpointName, handler, authorizationRequired, isDefault) {
+ServerApp.registerEndpoint = function (endpointName, handler, authorizationRequired, isDefault) {
   if (!appBinding.endpoints[endpointName]) {
     appBinding.endpoints[endpointName] = handler
     return appBinding.registerEndpoint(
@@ -232,27 +285,27 @@ App.registerEndpoint = function (endpointName, handler, authorizationRequired, i
  * @param {String} endpointName
  * @param {String} roleCode
  * @return {boolean} true if endpoint exists and role not already granted, false otherwise
- * @memberOf App
+ * @memberOf ServerApp
  */
-App.grantEndpointToRole = function (endpointName, roleCode) {
+ServerApp.grantEndpointToRole = function (endpointName, roleCode) {
   return appBinding.grantEndpointToRole(endpointName, roleCode)
 }
 
 /**
  * @method addAppLevelMethod
- * @deprecated Use {@link class:App.registerEndpoint App.registerEndpoint} instead
- * @memberOf App
+ * @deprecated Use {@link class:ServerApp.registerEndpoint App.registerEndpoint} instead
+ * @memberOf ServerApp
  */
-App.addAppLevelMethod = function () {
+ServerApp.addAppLevelMethod = function () {
   throw new Error('App.addAppLevelMethod is obsolete. Use App.registerEndpoint instead')
 }
 
 /**
  * @method  serviceMethodByPassAuthentication
  * @deprecated Use {@link class:App.registerEndpoint App.registerEndpoint} instead
- * @memberOf App
+ * @memberOf ServerApp
  */
-App.serviceMethodByPassAuthentication = function () {
+ServerApp.serviceMethodByPassAuthentication = function () {
   throw new Error('App.serviceMethodByPassAuthentication is obsolete. Use App.registerEndpoint instead')
 }
 
@@ -271,9 +324,9 @@ App.serviceMethodByPassAuthentication = function () {
  * @property {Object} uiSettings Section `uiSettings` of ubConfig
  * @property {Object} security
  */
-App.serverConfig = undefined
+ServerApp.serverConfig = undefined
 try {
-  App.serverConfig = argv.getServerConfiguration()
+  ServerApp.serverConfig = argv.getServerConfiguration()
 } catch (e) {
   console.error(e)
 }
@@ -282,18 +335,18 @@ try {
  * Application `package.json` content (parsed)
  * @type {Object}
  */
-App.package = require(path.join(process.configPath, 'package.json'))
+ServerApp.package = require(path.join(process.configPath, 'package.json'))
 
 /**
  * Full path to application static folder if any, '' if static folder not set
  * @type {String}
  * @readonly
  */
-App.staticPath = ''
-if (App.serverConfig.httpServer && App.serverConfig.httpServer['inetPub'] &&
-  App.serverConfig.httpServer['inetPub'].trim()) {
-  let sp = App.serverConfig.httpServer['inetPub']
-  App.staticPath = path.isAbsolute(sp)
+ServerApp.staticPath = ''
+if (ServerApp.serverConfig.httpServer && ServerApp.serverConfig.httpServer['inetPub'] &&
+  ServerApp.serverConfig.httpServer['inetPub'].trim()) {
+  let sp = ServerApp.serverConfig.httpServer['inetPub']
+  ServerApp.staticPath = path.isAbsolute(sp)
     ? sp
     : path.join(process.configPath, sp)
 }
@@ -303,18 +356,18 @@ if (App.serverConfig.httpServer && App.serverConfig.httpServer['inetPub'] &&
  * @type {String}
  * @readonly
  */
-App.defaultLang = App.serverConfig.application.defaultLang
+ServerApp.defaultLang = ServerApp.serverConfig.application.defaultLang
 
 /**
  * Custom settings for application from ubConfig.app.customSettings
  * @deprecated Use App.serverConfig.application.customSettings: Object instead
  * @type {String}
  */
-Object.defineProperty(App, 'customSettings', {
+Object.defineProperty(ServerApp, 'customSettings', {
   enumerable: true,
   get: function () {
     console.warn('App.customSettings is deprecated. Use App.serverConfig.application.customSettings instead')
-    return JSON.stringify(App.serverConfig.application.customSettings)
+    return JSON.stringify(ServerApp.serverConfig.application.customSettings)
   }
 })
 
@@ -323,9 +376,9 @@ Object.defineProperty(App, 'customSettings', {
  * @deprecated Use App.serverConfig.uiSettings: Object instead
  * @return {string}
  */
-App.getUISettings = function () {
+ServerApp.getUISettings = function () {
   console.warn('App.getUISettings is deprecated. Use App.serverConfig.uiSettings: Object instead')
-  return JSON.stringify(App.serverConfig.uiSettings)
+  return JSON.stringify(ServerApp.serverConfig.uiSettings)
 }
 
 /**
@@ -333,26 +386,26 @@ App.getUISettings = function () {
  * @type {String}
  * @readonly
  */
-App.serverURL = argv.serverURLFromConfig(App.serverConfig)
+ServerApp.serverURL = argv.serverURLFromConfig(ServerApp.serverConfig)
 
 /**
  * URL that the User from the internet will use to access your server. To be used in case server is behind a reverse proxy
  * @type {String}
  * @readonly
  */
-App.externalURL = App.serverConfig.httpServer.externalURL || App.serverURL
+ServerApp.externalURL = ServerApp.serverConfig.httpServer.externalURL || ServerApp.serverURL
 
 /**
  * List of a local server IP addresses CRLF (or CR for non-windows) separated
  */
-App.localIPs = _App.localIPs
+ServerApp.localIPs = _App.localIPs
 
 /**
  * Current application Domain
  * @deprecated UB >=4 use a App.domainInfo - a pure JS domain representation
  * @readonly
  */
-Object.defineProperty(App, 'domain', {
+Object.defineProperty(ServerApp, 'domain', {
   enumerable: true,
   get: function () {
     throw new Error('App.domain is obsolete. Use App.domainInfo')
@@ -363,7 +416,7 @@ Object.defineProperty(App, 'domain', {
  * For UB EE return true in case product license is exceed. For UB Se always `false`
  * @type {String}
  */
-Object.defineProperty(App, 'isLicenseExceed', {
+Object.defineProperty(ServerApp, 'isLicenseExceed', {
   enumerable: true,
   get: function () {
     return typeof appBinding.isLicenseExceed === 'function'
@@ -376,10 +429,10 @@ const getDomainInfo = appBinding.getDomainInfo
 let _domainCache
 /**
  * Extended information about application domain (metadata)
- * @memberOf App
+ * @memberOf ServerApp
  * @member {UBDomain} domainInfo
  */
-Object.defineProperty(App, 'domainInfo', {
+Object.defineProperty(ServerApp, 'domainInfo', {
   enumerable: true,
   get: function () {
     if (!_domainCache) {
@@ -397,7 +450,7 @@ Object.defineProperty(App, 'domainInfo', {
  * @param {String} key Key to retrive
  * @return {String}
  */
-App.globalCacheGet = function (key) {
+ServerApp.globalCacheGet = function (key) {
   return _App.globalCacheGet(key)
 }
 /**
@@ -405,7 +458,7 @@ App.globalCacheGet = function (key) {
  * @param {String} key  Key to put into
  * @param {String|null} value Value to put into this key. If === null then key will be remover from cache
  */
-App.globalCachePut = function (key, value) {
+ServerApp.globalCachePut = function (key, value) {
   _App.globalCachePut(key, value)
 }
 
@@ -414,7 +467,7 @@ App.globalCachePut = function (key, value) {
  * @param {String} entityName
  * @param {Number} instanceID
  */
-App.deleteFromFTSIndex = function (entityName, instanceID) {
+ServerApp.deleteFromFTSIndex = function (entityName, instanceID) {
   _App.deleteFromFTSIndex(entityName, instanceID)
 }
 /**
@@ -424,7 +477,7 @@ App.deleteFromFTSIndex = function (entityName, instanceID) {
  * @param {String} entityName
  * @param {Number} instanceID
  */
-App.updateFTSIndex = function (entityName, instanceID) {
+ServerApp.updateFTSIndex = function (entityName, instanceID) {
   _App.updateFTSIndex(entityName, instanceID)
 }
 
@@ -432,14 +485,14 @@ App.updateFTSIndex = function (entityName, instanceID) {
  * Databases connections pool
  * @type {Object<string, DBConnection>}
  */
-App.dbConnections = createDBConnectionPool(App.domainInfo.connections)
+ServerApp.dbConnections = createDBConnectionPool(ServerApp.domainInfo.connections)
 
 /**
  * Check database are used in current endpoint context and DB transaction is already active.
  * @param {String} connectionName
  * @return {Boolean}
  */
-App.dbInTransaction = function (connectionName) {
+ServerApp.dbInTransaction = function (connectionName) {
   return _App.dbInTransaction(connectionName)
 }
 /**
@@ -449,7 +502,7 @@ App.dbInTransaction = function (connectionName) {
  * @param {String} [connectionName]
  * @return {Boolean}
  */
-App.dbCommit = function (connectionName) {
+ServerApp.dbCommit = function (connectionName) {
   return connectionName ? _App.dbCommit(connectionName) : _App.dbCommit()
 }
 /**
@@ -459,7 +512,7 @@ App.dbCommit = function (connectionName) {
  * @param {String} [connectionName]
  * @return {Boolean}
  */
-App.dbRollback = function (connectionName) {
+ServerApp.dbRollback = function (connectionName) {
   return connectionName ? _App.dbRollback(connectionName) : _App.dbRollback()
 }
 /**
@@ -473,7 +526,7 @@ App.dbRollback = function (connectionName) {
  * @param {String} connectionName
  * @return {Boolean}
  */
-App.dbStartTransaction = function (connectionName) {
+ServerApp.dbStartTransaction = function (connectionName) {
   return _App.dbStartTransaction(connectionName)
 }
 
@@ -486,19 +539,19 @@ App.dbStartTransaction = function (connectionName) {
  * @param {boolean} doSetOutCookie If true set a out authorization cookie on success response (Negotiate only)
  * @return {Boolean}
  */
-App.authFromRequest = function (noHTTPBodyInResp = false, doSetOutCookie = false) {
+ServerApp.authFromRequest = function (noHTTPBodyInResp = false, doSetOutCookie = false) {
   return _App.authFromRequest(noHTTPBodyInResp, doSetOutCookie)
 }
 /**
  * Logout all users with the same name as current user, except currently logged user
  */
-App.logoutAllWithTheSameNameExceptMe = function () {
+ServerApp.logoutAllWithTheSameNameExceptMe = function () {
   return _App.logoutAllWithTheSameNameExceptMe()
 }
 /**
  * Logout a current user (kill current session)
  */
-App.logout = function () {
+ServerApp.logout = function () {
   return _App.logout()
 }
 /**
@@ -511,7 +564,7 @@ if App.els('uba_user', 'insert'){
  * @param {String} methodCode
  * @return {boolean}
  */
-App.els = function (entityCode, methodCode) {
+ServerApp.els = function (entityCode, methodCode) {
   return _App.els(entityCode, methodCode)
 }
 
@@ -543,7 +596,7 @@ function concurrentFileAccess() {
  * @param {string} csName Critical section name
  * @return {number}
  */
-App.registerCriticalSection = appBinding.registerCriticalSection
+ServerApp.registerCriticalSection = appBinding.registerCriticalSection
 /**
  * Waits for ownership of the specified critical section object. The function returns when the calling thread is granted ownership.
  *
@@ -552,13 +605,13 @@ App.registerCriticalSection = appBinding.registerCriticalSection
  * @method
  * @param {number} csIndex A critical section index returned by `App.registerCriticalSection`
  */
-App.enterCriticalSection = appBinding.enterCriticalSection
+ServerApp.enterCriticalSection = appBinding.enterCriticalSection
 /**
  * Releases ownership of the specified critical section
  * @method
  * @param {number} csIndex
  */
-App.leaveCriticalSection = appBinding.leaveCriticalSection
+ServerApp.leaveCriticalSection = appBinding.leaveCriticalSection
 
 /**
  * Enter a log recursion call.
@@ -579,12 +632,12 @@ App.leaveCriticalSection = appBinding.leaveCriticalSection
  * @method
  * @param {string} methodName
  */
-App.logEnter = appBinding.logEnter
+ServerApp.logEnter = appBinding.logEnter
 /**
  * Exit a log recursion call
  * @method
  */
-App.logLeave = appBinding.logLeave
+ServerApp.logLeave = appBinding.logLeave
 
 /**
  * Remove all user sessions (logout user).
@@ -592,14 +645,14 @@ App.logLeave = appBinding.logLeave
  * @param {number} userID
  * @return {boolean} true if user had had any session
  */
-App.removeUserSessions = appBinding.removeUserSessions || function(){}
+ServerApp.removeUserSessions = appBinding.removeUserSessions || function () {}
 
 /**
  * Is event emitter enabled for App singleton. Default is `false`
  * @deprecated Starting from 1.11 this property ignored (always TRUE)
  * @type {Boolean}
  */
-App.emitterEnabled = true
+ServerApp.emitterEnabled = true
 
 /**
  * Defense edition only,
@@ -610,7 +663,7 @@ App.emitterEnabled = true
  *
  * @type {string}
  */
-App.serverPublicCert = _App.serverPublicCert
+ServerApp.serverPublicCert = _App.serverPublicCert
 
 /**
  * BLOB stores methods. For usage examples see:
@@ -620,7 +673,7 @@ App.serverPublicCert = _App.serverPublicCert
  *  - {@link module:@unitybase/blob-stores~markRevisionAsPermanent App.blobStores.markRevisionAsPermanent} - mark specified revision of a historical store as permanent
  *  - {@link module:@unitybase/blob-stores~internalWriteDocumentToResp App.blobStores.internalWriteDocumentToResp} - mark specified revision of a historical store as permanent
  */
-App.blobStores = {
+ServerApp.blobStores = {
   getContent: blobStores.getContent,
   getContentPath: blobStores.getContentPath,
   putContent: blobStores.putContent,
@@ -638,6 +691,6 @@ App.blobStores = {
  * @type {Object}
  * @since UB@5.17.9
  */
-App.endpointContext = {}
+ServerApp.endpointContext = {}
 
-module.exports = App
+module.exports = ServerApp