+++ packages/adminui-vue/CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-- `processing.js`: `refresh` method emit `UB.connection.emit(`${masterEntityName}:refresh`, state.data.ID)` event
+ - `processing.js`: `refresh` method emits a `UB.connection.emit(`${masterEntityName}:refresh`, {ID: state.data.ID})` event.
+
 ### Changed
 
 ### Deprecated